coreos.unique.boot.failure fixed for aarch64
coreos.ignition.failure has been failing on RHCOS in general.

After failing to reproduce it locally I am removing it
from denylist in order to catch errors in the pipeline.
The goal of this PR is to allow it in RHCOS 4.16, catch the
errors from the log and re-introduce `coreos.ignition.failure`
to kola-denylist while working on solution.

See: https://github.com/coreos/coreos-assembler/issues/3670 and https://github.com/coreos/coreos-assembler/issues/3669